fix umbrella source paths in report
Ordering sub app paths to force the longest one to be matched first.
Also grabbing the apps path from the mix config so it will be appended
when building the stats path.

Paths to the source files in subapps would not properly reflect their
location in the directory. If a subapp's name contained another it was
possible for it to not have it's files listed correctly.

Given a project structure similar to:

```
umbrella/
    apps/
        myapp
        myapp_web
```

Where `myapp_web` has a dependency on `myapp` causing it to appear
after `myapp` in the results of `Mix.Dep.Umbrella.loaded`. This causes
all file paths reported to begin with `myapp`. Additionally the
`apps/` directory would not be prepended to allow coveralls.io to find
the source file.

The report would show these files:

```
myapp/lib/mod.ex
myapp/lib/web_mod.ex
```

Instead of:

```
apps/myapp/lib/mod.ex
apps/myapp_web/lib/web_mod.ex
```