Parallel support for separate CircleCI Workflow jobs
`CIRCLE_WORKFLOW_WORKSPACE_ID` is the same when "Rerun failed job" is
done while `CIRCLE_WORKFLOW_ID` changes, so use
`CIRCLE_WORKFLOW_WORKSPACE_ID` for `service_number`, so that the results from
the original and rerun are combined.  Use `CIRCLE_BUILD_NUM` for
`service_number` either in or out of Workflows.  This appears to match
the usage of `service_number` and `service_job_id` in
https://github.com/lemurheavy/coveralls-ruby/commit/cf20b3a3242483b7b5bf7fbdf149008aa0a20541,
which added parallel support for the older, parallel-in-1-job system to
coveralls-ruby - service_job_id varies over the parallel things while
service_number must be the same across all parallel siblings.