Remove `text-align: right` so filenames are easier to scan
When working on a large project with many deeply-nested files,
scanning the list of files for a particular one is difficult because
of the right-alignment. This change makes the file paths line up so
you can quickly tell when you're looking at a different folder. The
coverage percentage remains right-aligned, so that is still easy to
scan.