chore: PathReader.base_path config
This patch changes the `PathReader.base_path` function to try reading
from the `Application` first, using the current method as a fallback
mechanism in case the config is not available.

The reason for this change is that we want to instrument a running
elixir app while executing some acceptance tests and
`Mix.Project.config_files` does not have all the necessary information
to infer the path.