fix: use `File.cwd!/0` for fetching base path
Otherwise it will fail if tests use [`in_project/3`][1] with confusing
error about the fact that `nil` cannot be used in `Path.dirname/1`.

It should work 100% of the time as Mix will work only when ran in
directory with `mix.exs` in it, and when running in umbrella it will
change directory to the sub project before running tasks.

Ref hauleth/mix_unused#25

[1]: https://hexdocs.pm/mix/Mix.Project.html#in_project/3