Ignore single next line
Allows to ignore the coverage for the following line of code,
by putting a new type of comment `# coveralls-ignore-next-line`.

Ignoring a single line is a common use case, which currently can be done by wrapping the code
between the "start" and "stop" comments. If we can achieve the same with just one comment,
this produces less visual noise and leaves more space for the useful code.

Resolves #219